Production check: 45 adversarially verified fixes across 21 files. All 624 tests pass.
Key fixes:
- lvm_kriging mean_data shape validation (prevents buffer over-read)
- pseudo_gaussian_transform deterministic RandomState(seed)
- OMP critical section batching every 1000 iterations (~30-50% speedup)
- set_thread_num range validation
- Path traversal detection improvements (pre-normpath split, basedir param)
- SaveGSLIB uses validated path (closes TOCTOU window)
- C++ null checks in api.cpp, api_helpers.cpp, buffer_accessor_3d.h
- covariance_field.h int overflow prevention via long long cast
- select.h negative index guard (template-compatible)
- indicator_kriging.h thread-private probs vector, schedule(dynamic)
- my_kriging_weights.h unused var cleanup, single dpotrs_ call for OK
- sugarbox_neighbour_lookup.h uncommented bounds guard
- CMakeLists.txt BS_EXPORTING_PLUGIN, MKL link order fix
- hpgl.vcxproj /Gm removal, buffer_accessor_3d.h added, warning sync
- conftest.py fixture isolation (RandomState instead of global seed)
- test_kriging_complete.py nugget==sill edge case, range tests
- test_simulation_complete.py serial-vs-parallel statistical comparisons
- Deprecated test_kriging.py and test_simulation.py with warnings

Pipeline: 55 agents, 5 stages (discovery → adversarial verification →
fixes → post-fix review → adversarial verification). 64 findings →
45 actionable → all fixed → 13 post-fix findings → all LOW after
adversarial verification.

+++ src/geo_bsd/hpgl/covariance_field.h
@@ -153,13 +153,16 @@ public:
 	}
 };
 
-double
+	double
 covariance_field_t::value(
 	int x,
 	int y,
 	int z)const
 {
-	return m_data[vr_to_dec(m_ydiameter, m_xdiameter, z + m_zradius, y + m_yradius, x + m_xradius)];
+	return m_data[static_cast<size_t>(vr_to_dec(m_ydiameter, m_xdiameter,
+		static_cast<long long>(z) + m_zradius,
+		static_cast<long long>(y) + m_yradius,
+		static_cast<long long>(x) + m_xradius))];
 	/*int xx = x + m_xradius; 
 	int yy = y + m_yradius;
 	int zz = z + m_zradius;
